Split gRPC conversions into submodules analogous to modules in the types crates
