+++ mullvad-cli/src/cmds/bridge.rs
@@ -277,20 +277,25 @@ impl Bridge {
                 }
                 if let Some(new_providers) = providers {
                     constraints.providers =
-                        types::try_providers_constraint_from_proto(&new_providers).unwrap();
+                        types::relay_constraints::try_providers_constraint_from_proto(
+                            &new_providers,
+                        )
+                        .unwrap();
                 }
                 if let Some(new_ownership) = ownership {
-                    constraints.ownership = types::ownership_constraint_from_proto(new_ownership);
+                    constraints.ownership =
+                        types::relay_constraints::ownership_constraint_from_proto(new_ownership);
                 }
                 constraints
             }
             _ => {
                 let location = Constraint::<LocationConstraint>::from(location.unwrap_or_default());
-                let providers =
-                    types::try_providers_constraint_from_proto(&providers.unwrap_or_default())
-                        .unwrap();
+                let providers = types::relay_constraints::try_providers_constraint_from_proto(
+                    &providers.unwrap_or_default(),
+                )
+                .unwrap();
                 let ownership = ownership
-                    .map(types::ownership_constraint_from_proto)
+                    .map(types::relay_constraints::ownership_constraint_from_proto)
                     .unwrap_or_default();
 
                 BridgeConstraints {
